Moved test_prove_rs to test_integration.py
