+++ kmir/src/tests/integration/test_integration.py
@@ -424,3 +424,20 @@ def test_prove(spec: Path, tmp_path: Path, kmir: KMIR) -> None:
     for label in claim_labels:
         proof = Proof.read_proof_data(proof_dir, label)
         assert proof.passed
+
+PROVING_DIR = (Path(__file__).parent / 'data' / 'prove-rs').resolve(strict=True)
+PROVING_FILES = list(PROVING_DIR.glob('*.rs'))
+
+
+@pytest.mark.parametrize(
+    'rs_file',
+    PROVING_FILES,
+    ids=[spec.stem for spec in PROVING_FILES],
+)
+def test_prove_rs(rs_file: Path, kmir: KMIR) -> None:
+    should_fail = rs_file.stem.endswith('fail')
+    apr_proof = kmir.prove_rs(rs_file)
+    if not should_fail:
+        assert apr_proof.passed
+    else:
+        assert apr_proof.failed
